Lsapp update

+++ app/Http/Controllers/PostsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 
 class PostsController extends Controller
@@ -55,13 +56,34 @@ class PostsController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
 
-        //Create Post
+        // ? Handle file upload
+        if($request->hasFile('cover_image'))
+        {
+            // ? Get file name w/ extension
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // ? Get just file name
+            $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            // ? Get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // ? File name to store
+            $fileNameToStore = $fileName.'_'.time().'.'.$extension;
+            // ? Upload the image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
+        else 
+        {
+            $fileNameToStore = 'noimage.jpg';
+        }
+
+        // ? Create Post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post created');
@@ -88,6 +110,13 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // ? Check for correct user
+        if(auth()->user()->id !== $post->user_id)
+        {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -105,10 +134,29 @@ class PostsController extends Controller
             'body' => 'required',
         ]);
 
-        //Update Post
+        // ? Handle file upload
+        if ($request->hasFile('cover_image')) 
+        {
+            // ? Get file name w/ extension
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // ? Get just file name
+            $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            // ? Get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // ? File name to store
+            $fileNameToStore = $fileName . '_' . time() . '.' . $extension;
+            // ? Upload the image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
+
+        // ? Update Post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if ($request->hasFile('cover_image')) 
+        {
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts/')->with('success', $post->title . ' updated');
@@ -124,6 +172,18 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // ? Check for correct user
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+
+        if ($post->cover_image != 'noimage.jpg')
+        {
+            // ? Delete the image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+
         $post->delete();
 
         return redirect('/posts')->with('success', $post->title . ' removed');
